@blockstack-devops 2024-11-05
:tada: This PR is included in version 8.2.1 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@stacks/blockchain-api-client/v/8.2.1)
- [GitHub release](https://github.com/hirosystems/stacks-blockchain-api/releases/tag/v8.2.1)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: